move Product::$namePrefix and $nameSufix from project-base to the packages
- getFullname() renamed to getFullName()
