+++ packages/framework/src/Model/Product/Elasticsearch/ProductExportRepository.php
@@ -153,6 +153,8 @@ class ProductExportRepository
             ProductExportFieldProvider::PARTNO => $product->getPartno(),
             ProductExportFieldProvider::EAN => $product->getEan(),
             ProductExportFieldProvider::NAME => $product->getName($locale),
+            ProductExportFieldProvider::NAME_PREFIX => $product->getNamePrefix($locale),
+            ProductExportFieldProvider::NAME_SUFIX => $product->getNameSufix($locale),
             ProductExportFieldProvider::DESCRIPTION => $product->getDescription($domainId),
             ProductExportFieldProvider::SHORT_DESCRIPTION => $product->getShortDescription($domainId),
             ProductExportFieldProvider::BRAND => $product->getBrand() ? $product->getBrand()->getId() : '',
