+++ packages/framework/src/Model/Product/Elasticsearch/Scope/ProductExportFieldProvider.php
@@ -44,6 +44,8 @@ class ProductExportFieldProvider
     public const string HREFLANG_LINKS = 'hreflang_links';
     public const string PRODUCT_TYPE = 'product_type';
     public const string PRIORITY_BY_PRODUCT_TYPE = 'priority_by_product_type';
+    public const string NAME_PREFIX = 'name_prefix';
+    public const string NAME_SUFIX = 'name_sufix';
 
     /**
      * @return string[]
