+++ packages/frontend-api/src/Resources/config/graphql-types/ModelType/Product/ProductDecorator.types.yaml
@@ -10,6 +10,15 @@ ProductDecorator:
             name:
                 type: "String"
                 description: "Localized product name (domain dependent)"
+            namePrefix:
+                type: "String"
+                description: "Name prefix"
+            nameSuffix:
+                type: "String"
+                description: "Name suffix"
+            fullName:
+                type: "String!"
+                description: "The full name of the product, which consists of a prefix, name, and a suffix"
             shortDescription:
                 type: "String"
                 description: "Localized product short description (domain dependent)"
