+++ project-base/app/config/graphql/types/ModelType/Product/Product.types.yaml
@@ -22,15 +22,6 @@ Product:
             name:
                 type: "String!"
                 description: "Localized product name (domain dependent)"
-            namePrefix:
-                type: "String"
-                description: "Name prefix"
-            nameSuffix:
-                type: "String"
-                description: "Name suffix"
-            fullName:
-                type: "String!"
-                description: "The full name of the product, which consists of a prefix, name, and a suffix"
             storeAvailabilities:
                 type: "[StoreAvailability!]!"
                 description: "List of availabilities in individual stores"
