+++ project-base/app/src/Model/Product/Search/ProductElasticsearchConverter.php
@@ -15,8 +15,6 @@ class ProductElasticsearchConverter extends BaseProductElasticsearchConverter
     public function fillEmptyFields(array $product): array
     {
         $result = parent::fillEmptyFields($product);
-        $result['name_prefix'] = $product['name_prefix'] ?? null;
-        $result['name_sufix'] = $product['name_sufix'] ?? null;
         $result['usps'] = $product['usps'] ?? [];
         $result['availability_status'] = $product['availability_status'] ?? '';
         $result['product_available_stores_count_information'] = $product['product_available_stores_count_information'] ?? '';
