--- project-base/app/src/ProductFeed/Heureka/Model/FeedItem/HeurekaFeedItemFactory.php
@@ -1,46 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace App\ProductFeed\Heureka\Model\FeedItem;
-
-use Shopsys\FrameworkBundle\Component\Domain\Config\DomainConfig;
-use Shopsys\FrameworkBundle\Model\Product\Product;
-use Shopsys\ProductFeed\HeurekaBundle\Model\FeedItem\HeurekaFeedItem;
-use Shopsys\ProductFeed\HeurekaBundle\Model\FeedItem\HeurekaFeedItemFactory as BaseHeurekaFeedItemFactory;
-
-/**
- * @method string|null getBrandName(\App\Model\Product\Product $product)
- * @method \Shopsys\FrameworkBundle\Model\Pricing\Price getPrice(\App\Model\Product\Product $product, \Shopsys\FrameworkBundle\Component\Domain\Config\DomainConfig $domainConfig)
- * @method string|null getHeurekaCategoryFullName(\App\Model\Product\Product $product, \Shopsys\FrameworkBundle\Component\Domain\Config\DomainConfig $domainConfig)
- * @property \App\Model\Category\CategoryFacade $categoryFacade
- * @method __construct(\Shopsys\FrameworkBundle\Model\Product\Pricing\ProductPriceCalculationForCustomerUser $productPriceCalculationForCustomerUser, \Shopsys\ProductFeed\HeurekaBundle\Model\FeedItem\HeurekaProductDataBatchLoader $productDataBatchLoader, \Shopsys\ProductFeed\HeurekaBundle\Model\HeurekaCategory\HeurekaCategoryFacade $heurekaCategoryFacade, \App\Model\Category\CategoryFacade $categoryFacade, \Shopsys\FrameworkBundle\Model\Product\Availability\ProductAvailabilityFacade $productAvailabilityFacade, \Shopsys\FrameworkBundle\Component\Cache\InMemoryCache $inMemoryCache)
- */
-class HeurekaFeedItemFactory extends BaseHeurekaFeedItemFactory
-{
-    /**
-     * @param \App\Model\Product\Product $product
-     * @param \Shopsys\FrameworkBundle\Component\Domain\Config\DomainConfig $domainConfig
-     * @return \Shopsys\ProductFeed\HeurekaBundle\Model\FeedItem\HeurekaFeedItem
-     */
-    public function create(Product $product, DomainConfig $domainConfig): HeurekaFeedItem
-    {
-        $mainVariantId = $product->isVariant() ? $product->getMainVariant()->getId() : null;
-
-        return new HeurekaFeedItem(
-            $product->getId(),
-            $product->getFullname($domainConfig->getLocale()),
-            $this->productDataBatchLoader->getProductParametersByName($product, $domainConfig),
-            $this->productDataBatchLoader->getProductUrl($product, $domainConfig),
-            $this->getPrice($product, $domainConfig),
-            $mainVariantId,
-            $product->getDescriptionAsPlainText($domainConfig->getId()),
-            $this->productDataBatchLoader->getProductImageUrl($product, $domainConfig),
-            $this->getBrandName($product),
-            $product->getEan(),
-            $this->productAvailabilityFacade->getProductAvailabilityDaysForFeedsByDomainId($product, $domainConfig->getId()),
-            $this->getHeurekaCategoryFullName($product, $domainConfig),
-            $this->productDataBatchLoader->getProductCpc($product, $domainConfig),
-        );
-    }
-}
