+++ project-base/app/tests/FrontendApiBundle/Functional/Cart/MergeCartsTest.php
@@ -77,14 +77,14 @@ class MergeCartsTest extends GraphQlWithLoginTestCase
         $cartItems = $cart->getItems();
         self::assertCount(3, $cartItems);
 
-        self::assertEquals($anonymouslyAddedProduct->getFullname(), $cartItems[0]->getName(), 'First product name mismatch');
+        self::assertEquals($anonymouslyAddedProduct->getFullName(), $cartItems[0]->getName(), 'First product name mismatch');
         self::assertEquals($anonymouslyAddedProductQuantity, $cartItems[0]->getQuantity(), 'First product quantity mismatch');
 
-        self::assertEquals($anonymouslyAddedProduct2->getFullname(), $cartItems[1]->getName(), 'Second product name mismatch');
+        self::assertEquals($anonymouslyAddedProduct2->getFullName(), $cartItems[1]->getName(), 'Second product name mismatch');
         self::assertEquals(3, $cartItems[1]->getQuantity(), 'Second product quantity mismatch');
 
         $thirdProduct = $this->getReference(ProductDataFixture::PRODUCT_PREFIX . '72', Product::class);
-        self::assertEquals($thirdProduct->getFullname(), $cartItems[2]->getName(), 'Third product name mismatch');
+        self::assertEquals($thirdProduct->getFullName(), $cartItems[2]->getName(), 'Third product name mismatch');
         self::assertEquals(2, $cartItems[2]->getQuantity(), 'Third product quantity mismatch');
 
         $oldCart = $this->cartFacade->findCartByCartIdentifier($testCartUuid);
@@ -125,11 +125,11 @@ class MergeCartsTest extends GraphQlWithLoginTestCase
         self::assertCount(2, $cartItems);
 
         $firstProduct = $this->getReference(ProductDataFixture::PRODUCT_PREFIX . '1', Product::class);
-        self::assertEquals($firstProduct->getFullname(), $cartItems[0]->getName(), 'First product name mismatch');
+        self::assertEquals($firstProduct->getFullName(), $cartItems[0]->getName(), 'First product name mismatch');
         self::assertEquals(2, $cartItems[0]->getQuantity(), 'First product quantity mismatch');
 
         $secondProduct = $this->getReference(ProductDataFixture::PRODUCT_PREFIX . '72', Product::class);
-        self::assertEquals($secondProduct->getFullname(), $cartItems[1]->getName(), 'Second product name mismatch');
+        self::assertEquals($secondProduct->getFullName(), $cartItems[1]->getName(), 'Second product name mismatch');
         self::assertEquals(2, $cartItems[1]->getQuantity(), 'Second product quantity mismatch');
 
         $oldCart = $this->cartFacade->findCartByCartIdentifier($testCartUuid);
@@ -178,11 +178,11 @@ class MergeCartsTest extends GraphQlWithLoginTestCase
         self::assertCount(2, $cartItems);
 
         $secondProduct = $this->getReference(ProductDataFixture::PRODUCT_PREFIX . '1', Product::class);
-        self::assertEquals($secondProduct->getFullname(), $cartItems[0]->getName(), 'Second product name mismatch');
+        self::assertEquals($secondProduct->getFullName(), $cartItems[0]->getName(), 'Second product name mismatch');
         self::assertEquals(2, $cartItems[0]->getQuantity(), 'Second product quantity mismatch');
 
         $firstProduct = $this->getReference(ProductDataFixture::PRODUCT_PREFIX . '72', Product::class);
-        self::assertEquals($firstProduct->getFullname(), $cartItems[1]->getName(), 'Third product name mismatch');
+        self::assertEquals($firstProduct->getFullName(), $cartItems[1]->getName(), 'Third product name mismatch');
         self::assertEquals(2, $cartItems[1]->getQuantity(), 'Third product quantity mismatch');
 
         $oldCart = $this->cartFacade->findCartByCartIdentifier($testCartUuid);
